governance: add support for snapshot.org...
...and remove support for Harmony instance of snapshot. The API for
snapshot.org (according to snapshot.js) allows the creation / deletion
of proposals and voting on them. This draft commit only adds voting
support with others to follow.
