+++ pkg/governance/util.go
@@ -1,0 +1,14 @@
+package governance
+
+import (
+	"encoding/json"
+	"fmt"
+)
+
+func indent(v interface{}) string {
+	b, err := json.MarshalIndent(v, "", "  ")
+	if err != nil {
+		return fmt.Sprintf("%#v", v)
+	}
+	return string(b)
+}
